refactor: navigation related components moved to navigation directory
