+++ frontend/src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type {Metadata} from 'next'
 import {Inter} from 'next/font/google'
 import './globals.css'
-import NavBar from "@/app/components/NavBar";
+import NavBar from "@/app/components/navigation/NavBar";
 import Image from "next/image";
 
 const inter = Inter({subsets: ['latin']})
